test(cloudwatchlogs): group tests by behaviour, not by review round
Per review feedback: the rebase_negative / review_comment / review_findings /
rebase_seam files were named after the review round that produced them, which
does not mean anything to a future reader. Every test moved into the existing
behaviour-named file that owns that behaviour; no test was added or dropped
(99 before, 99 after).

pusher: batch_test (finalizers), circuitbreaker_test, pool_test, queue_test,
retry_test, retryheap_test, retryheap_recovery_test, sender_test,
poison_pill_test; shared stubs now in helpers_test.
cloudwatchlogs: new behaviour-named concurrency_test, shutdown_test, emf_test,
helpers_test replace rebase_seam / breaker_shutdown / emf_retry_header /
cloudwatchlogs_retryheap_target.

Also renamed TestRebaseSeam* to describe behaviour and stripped review-round
prefixes (M1/C1/Finding #N/Iteration-N/swarm) from doc comments, keeping the
behavioural explanation.
