+++ plugins/outputs/cloudwatchlogs/concurrency_test.go
@@ -6,6 +6,7 @@ package cloudwatchlogs
 import (
 	"net/http"
 	"net/http/httptest"
+	"strconv"
 	"strings"
 	"sync"
 	"sync/atomic"
@@ -13,6 +14,7 @@ import (
 	"time"
 
 	"github.com/influxdata/telegraf/testutil"
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
 	"github.com/aws/amazon-cloudwatch-agent/internal"
@@ -20,14 +22,78 @@ import (
 	"github.com/aws/amazon-cloudwatch-agent/tool/util"
 )
 
-type retryHeapTestEvent struct {
-	msg string
-	ts  time.Time
+// R2: the collapse rebase unioned #2190's sharedRetryer/sharedClient with the poison-pill
+// retryHeap/retryHeapProcessor in one struct. Both sets must be wired, and the shared pair
+// must exist even when concurrency is disabled -- it is not gated on the retry heap.
+func TestConcurrencyComponentsAllWired(t *testing.T) {
+	var puts atomic.Int32
+	srv := cwlServer(&puts, "")
+	defer srv.Close()
+
+	t.Run("ConcurrencyEnabled", func(t *testing.T) {
+		c := newPlugin(srv.URL, 4)
+		c.CreateDest("G", "S", -1, util.StandardLogGroupClass, nil)
+		// #2190 half
+		require.NotNil(t, c.sharedRetryer, "#2190 sharedRetryer lost in the rebase")
+		require.NotNil(t, c.sharedClient, "#2190 sharedClient lost in the rebase")
+		require.NotNil(t, c.targetManager)
+		// poison-pill half
+		require.NotNil(t, c.workerPool, "poison-pill workerPool lost in the rebase")
+		require.NotNil(t, c.retryHeap, "poison-pill retryHeap lost in the rebase")
+		require.NotNil(t, c.retryHeapProcessor, "poison-pill processor lost in the rebase")
+		// Not asserted: that the processor's retryer differs from the TargetManager's
+		// sharedRetryer. The processor does not expose its retryer, so there is nothing
+		// observable to compare -- a NotSame against the processor itself compares two
+		// different types and can never fail.
+		c.Close()
+	})
+
+	t.Run("ConcurrencyDisabled", func(t *testing.T) {
+		c := newPlugin(srv.URL, 1)
+		c.CreateDest("G", "S", -1, util.StandardLogGroupClass, nil)
+		// #2190 must still apply with the retry heap switched off
+		require.NotNil(t, c.sharedRetryer, "#2190 fix must not be gated on concurrency")
+		require.NotNil(t, c.sharedClient)
+		require.NotNil(t, c.targetManager)
+		assert.Nil(t, c.workerPool)
+		assert.Nil(t, c.retryHeap)
+		assert.Nil(t, c.retryHeapProcessor)
+		c.Close()
+	})
 }
 
-func (e *retryHeapTestEvent) Message() string { return e.msg }
-func (e *retryHeapTestEvent) Time() time.Time { return e.ts }
-func (e *retryHeapTestEvent) Done()           {}
+// P8: concurrency <= 1 leaves retryHeap nil and the sender falls back to the synchronous
+// path. main's default concurrency is 0, so this is the path existing installs run -- it must
+// deliver and shut down cleanly with no nil dereference from the poison-pill wiring.
+func TestSyncModeDefaultConcurrencyDelivers(t *testing.T) {
+	for _, conc := range []int{0, 1} {
+		t.Run("Concurrency"+strconv.Itoa(conc), func(t *testing.T) {
+			var puts atomic.Int32
+			srv := cwlServer(&puts, "")
+			defer srv.Close()
+
+			c := newPlugin(srv.URL, conc)
+			dest := c.CreateDest("G", "S", -1, util.StandardLogGroupClass, nil)
+			require.Nil(t, c.retryHeap, "retryHeap must be nil below the concurrency threshold")
+
+			var doneCount atomic.Int32
+			require.NoError(t, dest.Publish([]logs.LogEvent{
+				&countingEvent{msg: "sync-mode", ts: time.Now(), done: &doneCount},
+			}))
+
+			require.Eventually(t, func() bool { return doneCount.Load() == 1 }, 15*time.Second,
+				50*time.Millisecond, "synchronous path did not deliver (concurrency=%d)", conc)
+
+			closed := make(chan struct{})
+			go func() { c.Close(); close(closed) }()
+			select {
+			case <-closed:
+			case <-time.After(20 * time.Second):
+				t.Fatalf("Close() hung in synchronous mode (concurrency=%d)", conc)
+			}
+		})
+	}
+}
 
 // TestRetryHeapProcessorHasTargetManager guards the initialization order in getDest:
 // the RetryHeapProcessor captures targetManager by value, so targetManager must be built
