+++ plugins/outputs/cloudwatchlogs/emf_test.go
